Put complex utility functions inside FFT Swift file.
100% speedup (About 3x the first 1.2 build I tried).
