+++ GeekbenchSwift/fft.swift
@@ -1,8 +1,59 @@
 // Copyright (c) 2014 Primate Labs Inc.
 // Use of this source code is governed by the 2-clause BSD license that
 // can be found in the LICENSE file.
-
 import Foundation
+
+struct Complex : Printable {
+    internal var real : Float32
+    internal var imaginary : Float32
+    var description: String {
+        return String(format: "(%.5f, %.5fi)", real, imaginary)
+    }
+    
+    init() {
+        self.real = 0
+        self.imaginary = 0
+    }
+    
+    init(real : Float32, imaginary : Float32) {
+        self.real = real
+        self.imaginary = imaginary
+    }
+    
+    mutating func assign(real : Float32, imaginary : Float32) {
+        self.real = real
+        self.imaginary = imaginary
+    }
+    
+    mutating func assign(rhs : Complex) {
+        self.real = rhs.real
+        self.imaginary = rhs.imaginary
+    }
+    
+    mutating func add(rhs : Complex) {
+        self.real += rhs.real
+        self.imaginary += rhs.imaginary
+    }
+    
+    mutating func subtract(rhs : Complex) {
+        self.real -= rhs.real
+        self.imaginary -= rhs.imaginary
+    }
+}
+
+func * (left : Complex, right : Complex) -> Complex {
+    return Complex(real: left.real * right.real - left.imaginary * right.imaginary,
+        imaginary: left.real * right.imaginary + left.imaginary * right.real)
+}
+
+func + (left : Complex, right : Complex) -> Complex {
+    return Complex(real: left.real + right.real, imaginary: left.imaginary + right.imaginary)
+}
+
+func - (left : Complex, right : Complex) -> Complex {
+    return Complex(real: left.real - right.real, imaginary: left.imaginary - right.imaginary)
+}
+
 
 final class SFFTWorkload : Workload {
   let pi = Float32(acos(-1.0))
@@ -41,7 +92,8 @@ final class SFFTWorkload : Workload {
   }
 
   override func worker() {
-    for var chunkOrigin = 0; chunkOrigin < self.size; chunkOrigin += self.chunkSize {
+    //    for var chunkOrigin = 0; chunkOrigin < self.size; chunkOrigin += self.chunkSize {
+    for chunkOrigin in stride(from: 0, to: self.size, by: self.chunkSize) {
       reorderInputIntoOutput(chunkOrigin)
       executeInplaceFFTOnOutput(chunkOrigin)
     }
@@ -54,7 +106,8 @@ final class SFFTWorkload : Workload {
     // Right shift requred to account for unused leading zeros in the UInt32
     let shiftCorrection = countLeadingZeros(chunkSize) + 1
 
-    for var i : UInt32 = 0; i < chunkSize; ++i {
+    //for var i : UInt32 = 0; i < chunkSize; ++i {
+    for i in 0..<chunkSize {
       var o = i
       // Reverse the bits of o
       o = (o & 0x55555555) << 1 | (o & 0xAAAAAAAA) >> 1
